Changed type in benchmark (long to int).

+++ experiments/benchmark/benchmark.cpp
@@ -7,14 +7,14 @@
 
 typedef struct {
    int *ar;
-   long  n;
+   int  n;
 } subarray;
 
 void* incer(void *arg) {
    fprintf(stderr, "<target-app> thread worker fired!\n");
    fprintf(stderr, "<target-app> arg = %p (n = %d)\n", (subarray*) arg, ((subarray*) arg)->n);
 
-   long i = 0;
+   int i = 0;
    for (i = 0; i < ((subarray *)arg)->n; i++) {
       ((subarray *)arg)->ar[i]++;
    }
@@ -39,7 +39,7 @@ int main(void) {
    blocking_entry();
 
    int*        ar;
-   long         i;
+   int         i;
    pthread_t  th1;
    subarray   sb1;
